requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
ansible==8.7.0
ansible-core==2.15.9
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
ascii-magic==2.3.0
asttokens==2.4.1
async-timeout==4.0.3
asyncache==0.3.1
attrs==23.2.0
azure-core==1.30.0
azure-storage-blob==12.19.0
bcrypt==4.0.1
boto3==1.28.65
botocore==1.31.85
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.4
cycler==0.12.1
decorator==5.1.1
distlib==0.3.8
dnspython==2.6.1
docker==6.1.3
email-validator==2.1.0.post1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.103.2
filelock==3.13.1
fonttools==4.49.0
forbiddenfruit==0.1.4
frozenlist==1.4.1
gevent==23.9.1
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
h11==0.14.0
hagrid==0.3.111
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==5.0.7
ipython==8.18.1
isodate==0.6.1
itables==1.6.2
jax==0.4.20
jaxlib==0.4.20
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
kiwisolver==1.4.5
kr8s==0.13.1
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
multidict==6.0.5
names==0.3.0
networkx==2.8
numpy==1.24.4
opt-einsum==3.3.0
packaging==23.2
pandas==1.5.3
paramiko==3.4.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pycapnp==1.3.0
pycparser==2.21
pydantic==1.10.13
Pygments==2.17.2
pymongo==4.6.1
PyNaCl==1.5.0
pyOpenSSL==24.0.0
pyparsing==3.1.1
python-box==7.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jsonpath==0.10.3
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.1
redis==4.6.0
requests==2.31.0
resolvelib==1.0.1
RestrictedPython==7.0
result==0.10.0
rich==13.7.0
s3transfer==0.7.0
safetensors==0.4.1
scipy==1.12.0
sherlock==0.4.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
stack-data==0.6.3
starlette==0.27.0
syft==0.8.4
tqdm==4.66.1
traitlets==5.14.1
typeguard==2.13.3
typing_extensions==4.8.0
urllib3==1.26.18
uvicorn==0.24.0.post1
uvloop==0.19.0
virtualenv==20.25.1
virtualenv-api==2.1.18
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==12.0
yarl==1.9.4
zipp==3.17.0
zope.event==5.0
zope.interface==6.2
